Setup /favicon.ico as an alias
For example bots or simple script might not parse the html head section and expect the favicon at this exact location. Plus when debugging/testing with simple html markup not based on the templates, they might not even contain the path to the favicon leading to the browser requesting the file anyway.