Add support for image based subtitles to video live transcoding
We need to use the overlay filter to put the subtitle images on top of the video stream. We are using a fixed array for that, ffmpeg does support more subtitle codecs but I'm not sure which ones are image based right now, so the array can be extended in the future.