Transcode all audio and subtitle streams by default for LiveVideoTrans.
This way the user can choose the audio stream (e.g. different languages) in the frontend.
Subtitles still need to be configurable but for the current beta use case, having all at the same time is better. But there certaintly are **a lot** of use-cases where this behaviour would be unacceptable.