Make UserStorage use `number` instead of `bigint` for account ids
bigint feals like a third-class type in JS.
Use it in calculations 'n stuff? Yes.
But having it serialized into JSON and back? Tedious at best.

If we need more ids or ids that do not just iterate, we should store them as strings in JS and use something like a Twitter Snowflake