Replace `passport` lib for Login functionality with own implementation
It's finally gone \o/ - passport sounds nice in theory and might be nice for simple projects but it is uncomfortable to work with if you want additional functionality or something.

This implementation is far from great but it should allow for any OAuth2.0 provider to be added and work out-of-the-box.

The default OAuth providers in the config right now should be extracted into some kind of template system that allows using a template in the admin area so common OAuth2.0 providers do not need any reserach or setup by the admin other than setting up an app and provideing the client_id and client_secret values.