Adds ServerTimings header for better timings insight
This helps in identifiy culprits in slow requests.

It is enabled for now, but it should be disabled in production
in the future.

The `sessionEnd` is in place to make sure that
routes not setting timings do not just get their time
added to the `session` timing.
(Calling the stop method instead could lead to loss of some timings).