Clear File cache on write operations and fix tests
Caching broke the tests, partially because they do not have a
request and keep using the same instances
of the file-system-layer abstractions.

The tests probably lost meaning/impact/(?) and the file-system-abstraction
needs to be partially redesigned to better allow for 
caching that does not easily cause false-cache-hits.


Related to acd71580eda0ad0cd298fd9330723d14824c71ad