Implements first part of Drag'n Drop file-entries in the file list view
This is not done but it allows for drag'n drop of files
and it will swap their positions  on the client-side.
In the future this can be used to move files into a directory etc.

There is still a bug with a-Tags in the markup. Not correctly
detecting the target-Element and not putting borders around it.

It is also missing multi-select support yet.