Implement gstreamer live_transcode wrapper+move video.js to webpack setup
I hate the Webpack setup right now but I can use TypeScript
and use a bundled js with dependencies managed by npm.

ProcessBuilder seems to do something stupid with stdout
or something - Maybe related to logging.
Using it for the gstreamer_app process causes it to hang eventually.
That's why the child process is spawned manually for now.

A WebSocket connection now used to exchange information.
For now only the URI of the HLS manifest and total
duration is sent. But in the future the gstreamer app needs
to respond to stream selection (e.g. audio or subtitles)
and somewhat synchronize playback for that
(So gstreamer doesn't generate 20min of output while
the video player is paused and when the user returns
changes to a different stream causing a complete reencode).

The ApolloVideoPlayer now uses wrappers to allow
loading of custom stuff (like HLS) and fake some properties (like duration).
the live_transcode wrapper still needs to implement #seek.
The gstreamer app might have to seek and syncronize with the video player.