Recode/-structure how the frontend creates and manages player backends
Native, HLS, LiveTranscode, Twitch and YouTube are supported right now.
Some of them have a common base while stuff like
Twitch and YouTube have their own special logic.
This restructure of responsibilities makes it easier
to create and destroy different players.
The way we create them is still not great but for now *it works*.