Fix Video Player Controls and Cursor not shown/hidden properly
When the mouse is above the controls, they are not
supposed to be hidden. This happens because the element
has a 0px heigth. We are now checking based on the element
inside that has absolute position set (the reason why the parent is 0px).

Additionally the cursor is now being hidden too.