Completely re-implement Live-Transcode from gstreamer to ffmpeg
I deemed gstreamer not really suitable for this use-case. I learned a lot but
it comes with a lot of complexity in combination with this live_transcode feature.

ffmpeg does a lot of stuff automagically which helps **a lot**.
It can take care of the input part and with some (or a lot) of research about
the output format and encoders etc. you can get pretty decent results!

This implementation, like the previous, is still missing:
* Proper error handling/reporting to the user
* Seeking (restarting the process at a different position)
* Giving user control about which streams get encoded
* ...