Implements new video meta data editor prototype in Nuxt.js/Vue.js
Not happy with it... Will recode without Vue.js/Nuxt.js but it was a nice project
to get a feel for Vue.js and how it works.

But overall it feels like too much magic to feel realy nice and to enough magic to just work great.
I might not face the issues I face right now with some more workshops or tutorials,
but tbh. it already feels like I've put too much time into understanding Vue.js.