feat: Merge Movie and TV Show Directory analysers into one class
This allows for efficient analysis of a directory tree without having to
do all I/O operations essentially twice (once to check if the directory is a TV Show and again if it is a Movie).
This also makes sure that the identical logic for extra files (Trailers etc.) is reused.

Refactoring should be considered in the future to maybe split the TV Show
and Movie logic a bit more to make the class smaller and more readable