feat: Support deleting whole streams in the video meta editor
Right now writing the meta data changes and deleting the streams is split
into two seperate ffmped calls.
This is very slow but makes it a lot easier to map the stream indice values...
How the stream selection works in combination with the different meta data
features/mappings is not well enough documented and a bit confusing for me

To speed up the two writes, they check themselves if they are even necessary and skip themselves.