feat: partially implement media library sharing with other users
The /media/-endpoints mostly work as expected but the user
is unable to actually play/download the media.

All the file access always just assumes that the requested file
is owned by that user and stored at the default *virtual file system*.

So for this to properly work, file handling needs to be
mostly reworked to properly allow referencing other user's files.