fix(media): improve library scanner to only detect video files as videos
Before these changes, there are directory structures that
might have been incorrectly and result in nothing being indexed in that directory (recursively).

For example imagine the use-case of a DVD/BD that contains
multiple movies (some kind of collection).
And you create a directory for that DVD/BD and
sub directories for the individual movies (because they can exist alone and it is not a tv show/series).
But in addition to that, the parent dir you created for that DVD/BD
contains additional files that you might want to keep but
is not relevant to the indexing.
Such files would have been detected as 'potentions' and stop
the recursive search for more.