build(frontend): remove some files from the final build I don't want
`_dev/` contains assets that are only intended to be used in dev mode.
If something is using them at production, it is to be considered a bug.

I removed/empties shims.js too, as Apollo should be run with
Node.js v20 and newer. The file *shouldn't* do anything and it might
break the frontend in a future Svelte update,
but that file is also loaded in the context of the whole
Apollo application. I don't want these global states potentially messes with, if not required.