feat(frontend): implement video seeking functionality
I've added a fastSeek method that actually does a normal seek.
Firefox is the only browser that offers a fastSeek method but
to me it doesn't feel good when using fastSeek in combination with a normal seek at the end.

I've still introduced this method in preparation for the LiveTranscode.
This hopefully allows it to essentially fake seeking to a time
that is not generated yet and only initiate the costly operations when the actual seek is done.