feat: set the videos in the video player to always autoplay
The whole page is the video player. The user interaction
initiating the player to open, is enough I feel like,
for the user to expect it to just start playing. Why click twice?

Obviously, the browser might have a different opinion on that
and block the autoplay by-default. But that's on the user to allow/configure (which is good).