feat: synchronize additional video player state/data with clients (WIP)
This is kind of intended as a restore point for me.

With these changes some more state is synced with the
clients in preparation for the actual syncing done client-side.
I *think* all information for syncing the player is now
available client-side.

But we are still missing WebSocketMessage for state changes
not done by the reference player.
Right now, I am planning to solve this problem similar to
the previous/old/prototype implementation:
request a change, that is sent to the reference client,
and the reference client then performs the change.
But we'll see