fix: restore access to files etc. for player-session participants
By making the `sessionId` mandatory for most endpoints
and moving it into the path, it is a lot easier to know the
relevant `sessionId`.

That way, we also don't have to add the parameter at
every possible location (including the HLS manifest),
but can instead continue to use relative paths that will
automatically keep the `sessionId`.