feat: allow non-reference players/connections to (un-)pause and seek
This essentially uses the same idea used in the old/current
video player.
To make race conditions less likely/problematic, we have the
reference player perform the desired action.
This does mean that we need to wait for it to receive the request
and each client to receive the changed state.
But this way we have a clear chain of command and the
server does not have to do some 'heavy' lifting.