fix: cap live transcode HLS to the duration of the shortest stream
I encountered a bug where the duration differed between
the video and the two audio streams (all slightly different),
and the video player got stuck at the `waiting` event
2-3 seconds before the end of the longest stream.

With this, the output is capped to the shortest one,
which is the most easy to way to ensure equal length.
'Normal' input video should not cause any trouble with that I think.