feat: implement backend calling Svelte handler in prod and frontend proxying backend in dev
Now when developing:
start backend and frontend in dev mode but only access
the frontend's dev server. `/api/` calls are routed by vite
to the backend server. One exception for now are oRPC
requests from the server, which seems to have a bug in the
request handling or something...

In production:
We only start the backend server and it will import the
built SvelteKit frontend and call use its `handler.js`
to pass the request to it.