feat: support startOffset=auto in media PlayerSession's /start-watching
This looks up the currently stored watch progress from the
database and uses that as the start offsett.

This will allow the user interface to just start at that, instead
of relying on the offset that is used for rendering the UI at
that point in time.
This is one of the things I *hate* about YouTube. Yes, I
understand that it is different from what I am seeing,
but no I do not want to start that 4 hour live stream VOD
at 20 seconds because that browser tab with that
suggestion is a bit outdated >:c