feat: set up fastify to use zod for validation and serialization
There are two main reasons for doing this and choosing zod:
1. we already use zod because of oRPC and it would not
   support the libraries Fastify uses or mentions before
   zod in the documentation
   -> We use zod here too for consistency
      and to not add a new dependency
2. This allows for type-inference for route user inputs
   and keeps the validation schema and TypeScript types
   in sync. I very much prefer this over declaring both
   manually.

zod, or probably setting up any other library for this,
also makes specifying a JSON schema much nicer, smaller
and easier.

I also updated the validation error response to support
BigInt as part of the validation result.
This allows me to validate/serialize inputs as BigInt
and continue to provide useful information to the client.
(JSON#stringify does throw an Error for BigInt values)