feat: recode video thumbnail generation and implement an endpoint for it
The recode is partly based on the previous implementation,
but achieves better thumbnails overall in my testing, by
considering more metrics in the frame selection and
instructing ffmpeg to not just find keyframes but also use
the `scene` filter to find frames that are already different
from each other.

The request, without a cache hit, can take up to one second
in my testing.
That's a bit more than I would like, but I think the cache
makes it acceptable. We might pre-generate these images
in the future anyway, so they should not be generated
on-demand by a client request.

This implementation is still having problem with tv shows
that have the same opening etc. where consecutive
episodes might get the identical thumbnail/frame.
This is something that should be improved in the future,
but probably makes it more expensive (or needs to seek
a lot further into the video file which might be weird but
might also just be what we want? Realistically... is there
really a high risk for spoilers?)