feat: re-implement VideoAnalyser using ffprobe and zod for validation
Using zod allows us to trust in the ffprobe output structure
and make the code that relies on it easier to read and more stable.
This also introduces the potential to break for some
input files, because their output types for some values
differ from the schema. This is expected and acceptable
I think, because the code using these extracted values can
rest assured, that it has a specific type.